model_gateway/requirements.txt
# Model Gateway Requirements

# Web framework
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# AI Provider SDKs
anthropic>=0.42.0
boto3>=1.35.0
botocore>=1.35.0

# Configuration
pyyaml>=6.0.2
python-dotenv>=1.0.0

# Utilities
aiohttp>=3.11.0
